+++ src/app/(auth)/signup/page.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import { SignIn } from '@components/forms'
+import { SignUp } from '@components/forms'
 
 const Page = () => {
     return (
         <div className="mx-auto border p-4">
-            <SignIn />
+            <SignUp />
         </div>
     )
 }
