+++ src/app/@modal/(.)signup/page.tsx
@@ -1,12 +1,12 @@
 import React from 'react'
 import { Modal } from '@components/common'
-import { SignIn } from '@components/forms'
+import { SignUp } from '@components/forms'
 
 const Page = () => {
     return (
         <Modal>
             <div className="mx-auto border p-4">
-                <SignIn />
+                <SignUp />
             </div>
         </Modal>
     )
