+++ src/app/Dashboard/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from './components/Header'
+import { Header } from '@components/layout'
 
 const Dashboard = () => {
     return (
