+++ src/components/forms/index.ts
@@ -1,0 +1,4 @@
+import SignIn from './SignIn/signIn'
+import SignUp from './SignUp/signUp'
+
+export { SignIn, SignUp }
