+++ src/components/layout/index.ts
@@ -1,0 +1,3 @@
+import Header from './Header/Header'
+
+export { Header }
